+++ CMakeLists.txt
@@ -2350,6 +2350,7 @@ if(gRPC_INSTALL)
     LIBRARY DESTINATION ${gRPC_INSTALL_LIBDIR}
     ARCHIVE DESTINATION ${gRPC_INSTALL_LIBDIR}
   )
+
 endif()
 
 if(gRPC_BUILD_TESTS)
