+++ tools/doxygen/Doxyfile.c++.internal
@@ -1023,7 +1023,6 @@ include/grpcpp/impl/server_initializer.h \
 include/grpcpp/impl/server_initializer_impl.h \
 include/grpcpp/impl/service_type.h \
 include/grpcpp/resource_quota.h \
-include/grpcpp/resource_quota_impl.h \
 include/grpcpp/security/auth_context.h \
 include/grpcpp/security/auth_metadata_processor.h \
 include/grpcpp/security/auth_metadata_processor_impl.h \
